Visitor view is added.

+++ BIM313-Group5-FinalProject/Models/Visitor.cs
@@ -13,7 +13,7 @@ namespace BIM313_Group5_FinalProject.Models
         public string firstName { get; set; }
         public string lastName { get; set; }
         public string gender { get; set; }
-        public int age { get; set; }
+        public short age { get; set; }
         public string email { get; set; }
 
     }
